Test: migrate reportspec
