+++ test-tests/shared/src/main/scala/zio/test/TestMain.scala
@@ -8,7 +8,6 @@ object TestMain {
 
   def main(args: Array[String]): Unit = {
     val allTests: List[(String, AsyncBaseSpec)] = List(
-      ("DefaultTestReporterSpec", DefaultTestReporterSpec),
       ("SummaryBuilderSpec", SummaryBuilderSpec),
       ("GenSpec", GenSpec),
       ("SampleSpec", SampleSpec),
